+++ engine.h
@@ -1,7 +1,9 @@
 #ifndef ENGINE_H 
 #define ENGINE_H
 
-unsigned long long engine_perft(unsigned depth);
+#include <stdint.h>
+
+uint64_t engine_perft(unsigned depth);
 void engine_divide(int);
 void engine_test();
